+++ adapters/repos/db/vector/hnsw/search.go
@@ -32,15 +32,15 @@ func reasonableEfFromK(k int) int {
 	return ef
 }
 
-func (h *hnsw) SearchByID(id int, k int) ([]int, error) {
+func (h *hnsw) SearchByID(id int64, k int) ([]int64, error) {
 	return h.knnSearch(id, k, reasonableEfFromK(k))
 }
 
-func (h *hnsw) SearchByVector(vector []float32, k int, allowList helpers.AllowList) ([]int, error) {
+func (h *hnsw) SearchByVector(vector []float32, k int, allowList helpers.AllowList) ([]int64, error) {
 	return h.knnSearchByVector(vector, k, reasonableEfFromK(k), allowList)
 }
 
-func (h *hnsw) knnSearch(queryNodeID int, k int, ef int) ([]int, error) {
+func (h *hnsw) knnSearch(queryNodeID int64, k int, ef int) ([]int64, error) {
 	entryPointID := h.entryPointID
 	entryPointDistance, ok, err := h.distBetweenNodes(entryPointID, queryNodeID)
 	if err != nil || !ok {
@@ -48,7 +48,7 @@ func (h *hnsw) knnSearch(queryNodeID int, k int, ef int) ([]int, error) {
 	}
 
 	queryVector, err := h.vectorForID(context.Background(),
-		int32(queryNodeID))
+		int64(queryNodeID))
 	if err != nil {
 		return nil, errors.Wrapf(err, "could not get vector of object at docID %d", queryNodeID)
 	}
@@ -75,7 +75,7 @@ func (h *hnsw) knnSearch(queryNodeID int, k int, ef int) ([]int, error) {
 
 	flat := res.flattenInOrder()
 	size := min(len(flat), k)
-	out := make([]int, size)
+	out := make([]int64, size)
 	for i, elem := range flat {
 		if i >= size {
 			break
@@ -106,7 +106,7 @@ func (h *hnsw) searchLayerByVector(queryVector []float32,
 			return nil, errors.Wrapf(err, "calculate distance of current last result")
 		}
 
-		dist, ok, err := h.distanceToNode(distancer, int32(candidate.index))
+		dist, ok, err := h.distanceToNode(distancer, int64(candidate.index))
 		if err != nil {
 			return nil, errors.Wrap(err, "calculate distance between candidate and query")
 		}
@@ -147,10 +147,10 @@ func (h *hnsw) searchLayerByVector(queryVector []float32,
 	return results, nil
 }
 
-func newVisitedList(entrypoints binarySearchTreeGeneric) map[uint32]struct{} {
-	visited := map[uint32]struct{}{}
+func newVisitedList(entrypoints binarySearchTreeGeneric) map[uint64]struct{} {
+	visited := map[uint64]struct{}{}
 	for _, elem := range entrypoints.flattenInOrder() {
-		visited[uint32(elem.index)] = struct{}{}
+		visited[uint64(elem.index)] = struct{}{}
 	}
 	return visited
 }
@@ -165,7 +165,7 @@ func (h *hnsw) insertViableEntrypointsAsCandidatesAndResults(
 			// have an allow list (i.e. the user has probably set some sort of a
 			// filter restricting this search further. As a result we have to
 			// ignore items not on the list
-			if !allowList.Contains(uint32(ep.index)) {
+			if !allowList.Contains(uint64(ep.index)) {
 				continue
 			}
 		}
@@ -181,7 +181,7 @@ func (h *hnsw) insertViableEntrypointsAsCandidatesAndResults(
 func (h *hnsw) currentWorstResultDistance(results *binarySearchTreeGeneric,
 	distancer distancer.Distancer) (float32, error) {
 	if results.root != nil {
-		id := int32(results.maximum().index)
+		id := int64(results.maximum().index)
 		d, ok, err := h.distanceToNode(distancer, id)
 		if err != nil {
 			return 0, errors.Wrap(err,
@@ -202,8 +202,8 @@ func (h *hnsw) currentWorstResultDistance(results *binarySearchTreeGeneric,
 }
 
 func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
-	results *binarySearchTreeGeneric, connections []uint32,
-	visited map[uint32]struct{}, distancer distancer.Distancer, ef int,
+	results *binarySearchTreeGeneric, connections []uint64,
+	visited map[uint64]struct{}, distancer distancer.Distancer, ef int,
 	level int, allowList helpers.AllowList, worstResultDistance float32) error {
 	for _, neighborID := range connections {
 		if _, ok := visited[neighborID]; ok {
@@ -214,7 +214,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 		// make sure we never visit this neighbor again
 		visited[neighborID] = struct{}{}
 
-		distance, ok, err := h.distanceToNode(distancer, int32(neighborID))
+		distance, ok, err := h.distanceToNode(distancer, int64(neighborID))
 		if err != nil {
 			return errors.Wrap(err, "calculate distance between candidate and query")
 		}
@@ -226,7 +226,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 
 		resLenBefore := results.len() // calculating just once saves a bit of time
 		if distance < worstResultDistance || resLenBefore < ef {
-			candidates.insert(int(neighborID), distance)
+			candidates.insert(int64(neighborID), distance)
 			if level == 0 && allowList != nil {
 				// we are on the lowest level containing the actual candidates and we
 				// have an allow list (i.e. the user has probably set some sort of a
@@ -237,11 +237,11 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 				}
 			}
 
-			if h.hasTombstone(int(neighborID)) {
+			if h.hasTombstone(int64(neighborID)) {
 				continue
 			}
 
-			results.insert(int(neighborID), distance)
+			results.insert(int64(neighborID), distance)
 
 			// +1 because we have added one node size calculating the len
 			if resLenBefore+1 > ef {
@@ -255,7 +255,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 }
 
 func (h *hnsw) distanceToNode(distancer distancer.Distancer,
-	nodeID int32) (float32, bool, error) {
+	nodeID int64) (float32, bool, error) {
 	candidateVec, err := h.vectorForID(context.Background(), nodeID)
 	if err != nil {
 		var e storobj.ErrNotFound
@@ -279,14 +279,14 @@ func (h *hnsw) distanceToNode(distancer distancer.Distancer,
 // the underlying object seems to have been deleted, to recover from
 // this situation let's add a tombstone to the deleted object, so it
 // will be cleaned up and skip this candidate in the current search
-func (h *hnsw) handleDeletedNode(docID int32) {
-	if h.hasTombstone(int(docID)) {
+func (h *hnsw) handleDeletedNode(docID int64) {
+	if h.hasTombstone(int64(docID)) {
 		// nothing to do, this node already has a tombstone, it will be cleaned up
 		// in the next deletion cycle
 		return
 	}
 
-	h.addTombstone(int(docID))
+	h.addTombstone(int64(docID))
 	h.logger.WithField("action", "attach_tombstone_to_deleted_node").
 		WithField("node_id", docID).
 		Info("found a deleted node (%d) without a tombstone, "+
@@ -294,7 +294,7 @@ func (h *hnsw) handleDeletedNode(docID int32) {
 }
 
 func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
-	ef int, allowList helpers.AllowList) ([]int, error) {
+	ef int, allowList helpers.AllowList) ([]int64, error) {
 	if h.isEmpty() {
 		return nil, nil
 	}
@@ -333,7 +333,7 @@ func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
 
 	flat := res.flattenInOrder()
 	size := min(len(flat), k)
-	out := make([]int, size)
+	out := make([]int64, size)
 	for i, elem := range flat {
 		if i >= size {
 			break
@@ -345,32 +345,32 @@ func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
 }
 
 func (h *hnsw) selectNeighborsSimple(input binarySearchTreeGeneric,
-	max int, denyList helpers.AllowList) []uint32 {
+	max int, denyList helpers.AllowList) []uint64 {
 	flat := input.flattenInOrder()
 
 	maxSize := min(len(flat), max)
-	out := make([]uint32, maxSize)
+	out := make([]uint64, maxSize)
 	actualSize := 0
 	for i, elem := range flat {
-		if denyList != nil && denyList.Contains(uint32(elem.index)) {
+		if denyList != nil && denyList.Contains(uint64(elem.index)) {
 			continue
 		}
 
 		if i >= maxSize {
 			break
 		}
-		out[actualSize] = uint32(elem.index)
+		out[actualSize] = uint64(elem.index)
 		actualSize++
 	}
 
 	return out[:actualSize]
 }
 
-func (h *hnsw) selectNeighborsSimpleFromId(nodeId int, ids []uint32,
-	max int, denyList helpers.AllowList) ([]uint32, error) {
+func (h *hnsw) selectNeighborsSimpleFromId(nodeId int64, ids []uint64,
+	max int, denyList helpers.AllowList) ([]uint64, error) {
 	bst := &binarySearchTreeGeneric{}
 	for _, id := range ids {
-		dist, ok, err := h.distBetweenNodes(int(id), nodeId)
+		dist, ok, err := h.distBetweenNodes(int64(id), nodeId)
 		if err != nil {
 			return nil, errors.Wrap(err, "select neighbors simple from id")
 		}
@@ -379,7 +379,7 @@ func (h *hnsw) selectNeighborsSimpleFromId(nodeId int, ids []uint32,
 			// node was deleted in the underlying object store
 			continue
 		}
-		bst.insert(int(id), dist)
+		bst.insert(int64(id), dist)
 	}
 
 	return h.selectNeighborsSimple(*bst, max, denyList), nil
